+++ R/geocode.R
@@ -61,6 +61,7 @@ geocode <- function(addresses, autocomplete = FALSE, url_only = FALSE) {
   if (length(data) == 0) return(NULL)
 
   # Extract information
+  ids <- .get_ids(data)
   count <- 0
   geocode_failed <- character(0)
   geocoded <- data.table::rbindlist(
@@ -72,7 +73,7 @@ geocode <- function(addresses, autocomplete = FALSE, url_only = FALSE) {
         return(NULL)
       }
       result <- data.table::data.table(
-        id = count,
+        id = ids[count],
         address = df$Response$View$Result[[1]]$Location$Address$Label,
         street = df$Response$View$Result[[1]]$Location$Address$Street,
         houseNumber = df$Response$View$Result[[1]]$Location$Address$HouseNumber,
@@ -98,6 +99,7 @@ geocode <- function(addresses, autocomplete = FALSE, url_only = FALSE) {
 
   # Create sf, data.table, data.frame
   if (nrow(geocoded) > 0) {
+    rownames(geocoded) <- NULL
     return(
       sf::st_set_crs(
         sf::st_as_sf(geocoded, coords = c("lng", "lat")),