+++ NEWS.md
@@ -1,6 +1,7 @@
 # version 0.2.0.9000
 * Enhanced `traffic()`: Clarified that `from_dt` and `to_dt` have no effect on the traffic flow (`product = "flow"`). Traffic flow is always real-time. Detailed documentation of the variables in the return value.
 * Improved coverage of `testthat`.
+* Added an `"id"` column to the output of `geocode()` and removed the id ordered `row.names` in order to be consistent with other functions of the package. Using the `"id"` column the addresses to geocode can be joined to the coordinates after geocoding.
 
 # version 0.2.0