+++ NEWS.md
@@ -1,4 +1,4 @@
-# version 0.2.0.9000
+# version 0.2.1
 
 * Enhanced `traffic()`: Clarified that `from_dt` and `to_dt` have no effect on the traffic flow (`product = "flow"`). Traffic flow is always real-time. Detailed documentation of the variables in the return value.
 * Improved coverage of `testthat`.